requirement.txt
astroid==2.3.3
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
env-file==2019.4.13
filelock==3.0.12
get==2019.4.13
getenv==0.1.0
idna==2.8
importlib-metadata==1.3.0
isort==4.3.21
jwt==0.6.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
mysql-connector==2.2.9
numpy==1.18.0
packaging==19.2
pluggy==0.13.1
post==2019.4.13
public==2019.4.13
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.5
pytest==5.3.2
python-dotenv==0.10.3
python-env==1.0.0
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.13.0
toml==0.10.0
tox==3.14.2
typed-ast==1.4.0
urllib3==1.25.7
values==2019.4.13
virtualenv==16.7.9
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
